requirements.txt
fastapi
uvicorn
psycopg2-binary
neo4j
openai
requests
python-dotenv